+++ src/features/note/api/noteApi.ts
@@ -1,4 +1,4 @@
-import { type UpdateNoteRequest, type NoteListResponse, type NoteResponse } from "@/features/note/model";
+import { type NoteListResponse, type NoteResponse } from "@/features/note/model";
 
 export const getNotesByStudyGoalId = async (studyGoalId: number): Promise<NoteListResponse> => {
   const response = await fetch(`/api/notes/${studyGoalId}`, {
@@ -30,21 +30,3 @@ export const getNoteById = async (noteId: number): Promise<NoteResponse> => {
   return response.json();
 };
 
-export const updateNote = async (noteData: UpdateNoteRequest): Promise<NoteResponse> => {
-  const response = await fetch(`/api/notes/${noteData.id}`, {
-    method: 'PATCH',
-    headers: {
-      'Content-Type': 'application/json',
-    },
-    body: JSON.stringify({
-      title: noteData.title,
-      content: noteData.content,
-    }),
-  });
-
-  if (!response.ok) {
-    throw new Error('노트 수정에 실패했습니다.');
-  }
-
-  return response.json();
-};
