+++ src/features/note/edit-note/index.ts
@@ -1,0 +1,2 @@
+export { NoteEditor } from './ui/NoteEditor';
+export type { UpdateNoteRequest } from './model/types';
