+++ src/features/note/edit-note/model/types.ts
@@ -1,0 +1,5 @@
+export interface UpdateNoteRequest {
+  id: number;
+  title: string;
+  content: string;
+}
