+++ src/features/note/edit-note/ui/NoteEditor.tsx
