+++ src/features/note/model/index.ts
@@ -1,2 +1,2 @@
 export { noteKeys } from './queryKeys';
-export { type Note, type NoteListResponse, type NoteResponse, type UpdateNoteRequest } from './types';
+export { type Note, type NoteListResponse, type NoteResponse } from './types';
