+++ src/features/note/model/types.ts
@@ -1,24 +1,4 @@
-export interface Note {
-  id: number;
-  title: string;
-  content: string;
-  studyGoalId: number;
-  todoId: number;
-  todoTitle: string;
-  createdAt: string;
-  updatedAt: string;
-}
 
-export interface NoteListResponse {
-  notes: Note[];
-}
 
-export interface NoteResponse {
-  note: Note;
-}
 
-export interface UpdateNoteRequest {
-  id: number;
-  title: string;
-  content: string;
-} 
+
